+++ package.json
@@ -56,9 +56,10 @@
     "node": ">=14.16"
   },
   "dependencies": {
+    "@resvg/resvg-js": "2.6.2",
     "@types/chrome": "^0.1.42",
-    "colorthief": "3.3.1",
     "@types/web-app-manifest": "1.0.9",
+    "colorthief": "3.3.1",
     "esbuild": "0.27.4",
     "node-watch": "0.7.4",
     "open": "11.0.0",
